+++ src/Components/Core/Presentation/React/LearningSpaceDisplay/ProgressScorePanel/LevelUpModal/LevelUpModalViewModel.ts
@@ -1,0 +1,6 @@
+import Observable from "src/Lib/Observable";
+
+export default class LevelUpModalViewModel {
+  isOpen: Observable<boolean> = new Observable<boolean>(false);
+  timeToClose: number = 2500;
+}