+++ src/Components/CoreTest/Presentation/React/LearningSpaceDisplay/ProgressScorePanel/LevelUpModal/LevelUpModalController.test.ts
@@ -1,0 +1,18 @@
+import LevelUpModalController from "../../../../../../Core/Presentation/React/LearningSpaceDisplay/ProgressScorePanel/LevelUpModal/LevelUpModalController";
+import LevelUpModalViewModel from "../../../../../../Core/Presentation/React/LearningSpaceDisplay/ProgressScorePanel/LevelUpModal/LevelUpModalViewModel";
+
+describe("LevelUpModalController", () => {
+  let systemUnderTest: LevelUpModalController;
+  let viewModelMock: LevelUpModalViewModel;
+
+  beforeEach(() => {
+    viewModelMock = new LevelUpModalViewModel();
+  });
+
+  test("close sets isOpen in viewmodel to false", () => {
+    viewModelMock.isOpen.Value = true;
+    systemUnderTest = new LevelUpModalController(viewModelMock);
+    systemUnderTest.close();
+    expect(viewModelMock.isOpen.Value).toBeFalsy();
+  });
+});